+++ Programming/Nevo/app/src/main/java/com/medcorp/model/User.java
@@ -8,7 +8,7 @@ public class User {
 
     private int age = 18;
 
-    private int weight = 77;//kg
+    private float weight = 77;//kg
 
     private int height = 175;//cm
 
@@ -56,7 +56,7 @@ public class User {
         this.age = age;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(float weight) {
         this.weight = weight;
     }
 
@@ -80,7 +80,7 @@ public class User {
         return age;
     }
 
-    public int getWeight() {
+    public float getWeight() {
         return weight;
     }
